requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-documentai==3.3.0
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
rsa==4.9
urllib3==2.3.0
